+++ changelog.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [10.2.0] - Not Yet Released
+
+### Changed
+- Performed a number of refactoring (issue [#408](https://github.com/JasonBock/Rocks/issues/408))
+	- Improved the performance of `Verify()`
+
 ## [10.1.0] - 2026-02-24
 
 ### Added